requirements.txt
segno
Flask
gunicorn
python-dotenv